+++ packages/hyper_render_clipboard/lib/src/interfaces/file_system_provider.dart
@@ -1,8 +1,0 @@
-import 'dart:io';
-
-abstract class FileSystemProvider {
-  Future<Directory> getStorageDirectory();
-  Future<Directory> getCacheDirectory();
-  Future<Directory> getDocumentsDirectory();
-  Future<Directory?> getDownloadsDirectory();
-}